Model improvements

+++ avid/encoder.py
@@ -29,10 +29,12 @@ class SpeciesEmbed(nn.Module):
         mask: Bool[Array, ''],
         training: bool,
     ):
-        spec_embed = nn.Embed(self.n_species, self.species_embed_dim, name='species_embed')
+        spec_embed = nn.Embed(
+            self.n_species, self.species_embed_dim, name='species_embed', dtype=x.dtype
+        )
 
         input_embeds = jnp.concat([spec_embed(spec), x[..., None]], axis=-1)
-        return self.embed_module(input_embeds, training) * mask.astype(jnp.float32)
+        return self.embed_module(input_embeds, training) * mask.astype(x.dtype)
 
 
 class ReduceSpeciesEmbed(nn.Module):
